// client/models/AchievementDetailsAllOf.ts
/**
 * Odd Giants Public API
 * Odd Giants Public API
 *
 * OpenAPI spec version: 0.0.1
 * Contact: ustice@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Achievement } from './Achievement';
import { Item } from './Item';
import { Recipe } from './Recipe';
import { Upgrade } from './Upgrade';
import { HttpFile } from '../http/http';

export class AchievementDetailsAllOf {
    'connectionsFrom': Array<Achievement & any>;
    'connectionsTo': Array<Achievement & any>;
    'requiredForUpgrades': Array<Upgrade & any>;
    'requiredForRecipes': Array<Recipe & any>;
    'items': Array<Item & any>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "connectionsFrom",
            "baseName": "connections_from",
            "type": "Array<Achievement & any>",
            "format": ""
        },
        {
            "name": "connectionsTo",
            "baseName": "connections_to",
            "type": "Array<Achievement & any>",
            "format": ""
        },
        {
            "name": "requiredForUpgrades",
            "baseName": "required_for_upgrades",
            "type": "Array<Upgrade & any>",
            "format": ""
        },
        {
            "name": "requiredForRecipes",
            "baseName": "required_for_recipes",
            "type": "Array<Recipe & any>",
            "format": ""
        },
        {
            "name": "items",
            "baseName": "items",
            "type": "Array<Item & any>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AchievementDetailsAllOf.attributeTypeMap;
    }

    public constructor() {
    }
}