// client/models/AuthenticationCredentials.ts
/**
 * Odd Giants Public API
 * Odd Giants Public API
 *
 * OpenAPI spec version: 0.0.1
 * Contact: ustice@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class AuthenticationCredentials {
    'email': string;
    'password': string;
    'remember'?: boolean;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "email",
            "baseName": "email",
            "type": "string",
            "format": "email"
        },
        {
            "name": "password",
            "baseName": "password",
            "type": "string",
            "format": ""
        },
        {
            "name": "remember",
            "baseName": "remember",
            "type": "boolean",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AuthenticationCredentials.attributeTypeMap;
    }

    public constructor() {
    }
}