// client/models/CategoryAchievements.ts
/**
 * Odd Giants Public API
 * Odd Giants Public API
 *
 * OpenAPI spec version: 0.0.1
 * Contact: ustice@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Category } from './Category';
import { HttpFile } from '../http/http';

export class CategoryAchievements {
    'categories'?: Array<Category>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "categories",
            "baseName": "categories",
            "type": "Array<Category>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CategoryAchievements.attributeTypeMap;
    }

    public constructor() {
    }
}