// client/models/GiantDetailsAllOf.ts
/**
 * Odd Giants Public API
 * Odd Giants Public API
 *
 * OpenAPI spec version: 0.0.1
 * Contact: ustice@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { GiantId } from './GiantId';
import { Skill } from './Skill';
import { HttpFile } from '../http/http';

export class GiantDetailsAllOf {
    'skills': Array<Skill & any>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "skills",
            "baseName": "skills",
            "type": "Array<Skill & any>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return GiantDetailsAllOf.attributeTypeMap;
    }

    public constructor() {
    }
}